// apps/backend/src/facturacion/enums/tipo-comprobante.enum.ts
export enum TipoComprobante {
  factura_a = 'factura_a',
  factura_b = 'factura_b',
  factura_c = 'factura_c',
  nota_credito_a = 'nota_credito_a',
  nota_credito_b = 'nota_credito_b',
  nota_credito_c = 'nota_credito_c',
  remito = 'remito',
  presupuesto = 'presupuesto',
  ticket = 'ticket',
  recibo = 'recibo',
}